Silence deprecation warning for crypto:sha_mac/2
Erlang R16B deprecated the *_mac/2 familiy of functions in the crypto module.
In Erlang R15 the new hmac(*,...) functions are not yet available. So to
keep this module compatible with R16 _and_ R15 we silence the deprecation
warining. When R17ff actually removes the old functions, we should start
using the new functions and revert this commit.